update API Manager Export
