+++ lib/data/api_manager/api_manager.dart
@@ -1,0 +1,4 @@
+export 'api_manager_implementation.dart';
+export 'failure_handler.dart';
+export 'models/models.dart';
+export 'status_checker/status_checker.dart';
