+++ lib/data/api_manager/models/failures/failures.dart
@@ -1,0 +1,11 @@
+export 'connection_failure.dart';
+export 'error_failure.dart';
+export 'failure.dart';
+export 'failure_info.dart';
+export 'reportable_failure.dart';
+export 'request_canceled_failure.dart';
+export 'service_not_available_failure.dart';
+export 'session_ended_failure.dart';
+export 'type_failure.dart';
+export 'unkown_failure.dart';
+export 'validation_failure.dart';
